pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "empathy-research"
version = "0.1.0"
description = "Active Inference Research: Empathy, ToM, and Social Learning"
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "active inference",
    "theory of mind",
    "social learning",
    "empathy",
    "prisoner's dilemma",
    "multi-agent",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies (required by all)
dependencies = [
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
# Prisoners Dilemma + Graph Navigation (pymdp-based)
pymdp = [
    "inferactively-pymdp @ git+https://github.com/tverbele/pymdp.git@si_branch_observations",
    "networkx>=3.1.0",
    "pydot>=2.0.0",
]

# Clean Up task (numba-based, no pymdp)
clean-up = [
    "numba>=0.57.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]

# Development tools
dev = [
    "pytest>=7.3.0",
    "pytest-xdist>=3.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

# Full installation (all optional dependencies)
all = [
    "empathy-research[pymdp,clean-up,dev]",
]

[project.scripts]
run-clean-up = "empathy.clean_up.experiment.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["empathy*"]

[tool.setuptools.package-data]
"empathy.clean_up" = ["configs/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests", "src/empathy"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests", "scripts"]
